requirements.txt
requests==2.21.0
pandas==1.1.5
python-dotenv==0.21.0